
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_crime_reports.R
\encoding{UTF-8}
\name{clean_crime_reports}
\alias{clean_crime_reports}
\title{Crime Data Cleaning}
\usage{
clean_crime_reports(
  data,
  crime_var,
  date_var,
  date_format,
  time_period = "year",
  geography = NULL
)
}
\arguments{
\item{data}{The crime incident data to be aggregated}

\item{crime_var}{Categorical variable distinguishing different types of crime. This variable will be used for aggregating.}

\item{date_var}{Indicator representing the time of a given crime incident}

\item{date_format}{Vector indicating the format of dates. Uses lubricate date format. See documentation of lubridate for available options}

\item{time_period}{Time interval at which data will be aggregated. Available options are "year","year_month", and "year_quarter"}

\item{geography}{Specified geography for additional grouping if desired. If specified, data will nest time intervals within geographies.}
}
\value{
A grouped dataframe of counts with a column for each crime type in "crime_var"
}
\description{
Aggregates crime incident data by producing counts within certain time periods.
}
\examples{
aggregated_crime_data <- clean_crime_reports(raw_crime_data, crime_category,incident_date,"ymd_HS","year_quarter","bg")



}

