
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\encoding{UTF-8}
\name{distribution_bar}
\alias{distribution_bar}
\title{Plot distributions}
\usage{
distribution_bar(data, variable, value = "Count", threshold_value = 0)
}
\arguments{
\item{data}{Data used for plotting}

\item{variable}{The variable for which the distribution is to be plotted}

\item{threshold_value}{Only categories with a percentage of observations greater than the specified threshold will be plotted. Threshold takes the form of a}
}
\value{
a ggplot object
}
\description{
Plots the distribution of a specified categorical indicator
}

