
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_crime_map.R
\encoding{UTF-8}
\name{FC_crime_map}
\alias{FC_crime_map}
\title{Map Crime}
\usage{
FC_crime_map(
  data,
  indicator,
  underlay_counties = NULL,
  discrete = FALSE,
  roadmap = NULL,
  color_palette = "viridis",
  legend_lab = NULL
)
}
\arguments{
\item{data}{Shapefile containing the indicator to be mapped}

\item{indicator}{The variable within "data" whose values are to be mapped}

\item{discrete}{Is "indicator" discrete or continuous. Required for correct color palette specification}

\item{roadmap}{Adds major roads to the map. Input determines for which geographic area roads will be displayed. See osmdata documentation for available options.}

\item{color_palette}{Adjusts color palette. See scale_color_brewer for some available palettes.}

\item{legend_lab}{Adjusts legend label. Default is name of the indicator.}

\item{include_underlay}{Adds a layer of block groups for specified counties.}
}
\value{
A ggplot object
}
\description{
Maps values of a specified indicator across block groups
}

