
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_census_measures.R
\name{get_key_census_metric}
\alias{get_key_census_metric}
\title{Extract Key Census Measures}
\usage{
get_key_census_metric(geo_level, surv, state, counties = NULL, year = 2020)
}
\arguments{
\item{geo_level}{The geography level. See https://walker-data.com/tidycensus/articles/basic-usage.html for available levels}

\item{state}{The U.S. State containing the desired geography}

\item{counties}{The counties for which data is to be extracted. Default is NULL}

\item{year}{the year for which data is to be gathered. Default is 2020.}

\item{survey}{The survey used. Includes all American Community Surveys (1-year,3-year,etc.)}
}
\value{
A dataframe with key census indicators
}
\description{
Gathers a number of useful census measures. 
Note that this function leverages the 'get_acs' function from the 'tidycensus' package.
}
\details{
tidycensus relies on the US Census Bureau API to access Census data. 
In order to use tidycensus, you need to obtain an API key from the US Census Bureau and provide it when making requests for data. 
This API key is used to authenticate your access to the Census Bureau's data and services.

To set the API key run the following code: 'Sys.setenv(CENSUS_KEY = YOUR KEY)'
}

