
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\encoding{UTF-8}
\name{merge_files}
\alias{merge_files}
\title{Merge Data Files}
\usage{
merge_files(folder_path, file_type = ".csv", columns = NULL)
}
\arguments{
\item{columns}{Which columns from the data files to retain. Defaults to all columns.}

\item{folder}{The folder containing the .csv or .xlsx files}

\item{filetype}{The file type. Defaults to .csv. Options are ".csv" or ".xlsx".}
}
\value{
A tibble dataframe
}
\description{
Merges All CSV or Excel files Contained in Specified Folder
}

