% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_its.R
\encoding{UTF-8}
\name{plot_its}
\alias{plot_its}
\title{Plot Interrupted Time Series Trend}
\usage{
plot_its(
  data,
  time_var,
  outcome,
  treat_start,
  smooth = FALSE,
  xlab = "Time",
  ylab = "Value",
  treat_label_y = NULL
)
}
\arguments{
\item{data}{Data used for plotting}

\item{time_var}{Time variable. Can be expressed as an "integer" or "Date"}

\item{outcome}{Variable name of the outcome of interest}

\item{treat_start}{The start time of the treatment}

\item{smooth}{Whether to add a smoothed trend line, Default is NULL}

\item{xlab}{X-axis Title}

\item{ylab}{Y-axis Title}

\item{treat_label_y}{The heigh at which the "Treatment" label appears.}
}
\description{
Plots the time trend of an indicator before and after an intervention at time t
}
