
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\encoding{UTF-8}
\name{plot_timetrend}
\alias{plot_timetrend}
\title{Plot Time Trends}
\usage{
plot_timetrend(
  data,
  y_var,
  time_indicator,
  smoothing_method = NULL,
  linecolor = "black",
  linewidth = NULL,
  text_size = 14,
  theme = "minimal",
  xlab = "Time Period",
  ylab = "Value",
  group = NULL,
  group_colors = "Set1"
)
}
\arguments{
\item{data}{Data used for plotting}

\item{time_indicator}{The time indicator that will appear on the x-axis.}

\item{smoothing_method}{Smoothing method that will be passed to geom_smooth. See documentation for geom_smooth: https://ggplot2.tidyverse.org/reference/geom_smooth.html}

\item{linecolor}{The color of the trend line. Default is "black"}

\item{text_size}{The size of the text. Default is 14}

\item{theme}{The plot theme. See for available options: https://ggplot2.tidyverse.org/reference/ggtheme.html}

\item{xlab}{X-axis label}

\item{ylab}{Y-axis label}

\item{group}{Grouping variable. Default is null. Plots distinct time trends for each group if desired}

\item{group_colors}{Color palette used to distinguish each group.}

\item{linwidth}{The width of the trend line.}
}
\value{
A ggplot object
}
\description{
Plots the longitudinal trend of a specified indicator
}

